% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aoi.R
\name{map.aoi}
\alias{map.aoi}
\title{Map Research AOI}
\usage{
map.aoi(
  aoi,
  basemap = NULL,
  dem_raster = NULL,
  dem_z = 14,
  map_crs = NULL,
  map_ext = 1,
  watermask = TRUE,
  aoi_colour = "#ff0000",
  landscape = NULL,
  aspectratio = 16/9
)
}
\arguments{
\item{aoi}{Either sf-geometry, SpatVector or valid path to a file openable with `sf::st_read`. Should contain geometries representing an area of interest. Only mandatory argument.}

\item{basemap}{Either `NULL`, a SpatRaster or a valid path to a file openable with `terra::rast()`. Custom 3-band basemap used instead of one created automatically from a DEM.}

\item{dem_raster}{Either `NULL`, a SpatRaster or a valid path to a file openable with `terra::rast()`. Used to calculate a hillshade for the background if basemap is `NULL`. Downloads a DEM using `elevatr` if `NULL`.}

\item{dem_z}{Numeric, values between 0 and 14. Determines the zoom level of the dem downloaded if both `dem_raster` and `basemap` are `NULL`. More information at https://github.com/tilezen/joerd/blob/master/docs/data-sources.md#what-is-the-ground-resolution . Defaults to `14`.}

\item{map_crs}{Numeric or `NULL`. EPSG code of crs used in both processing and visualisation. If `NULL` the crs of the AOI is used. Defaults to `NULL`.}

\item{map_ext}{Either numeric, sf-geometry, SpatVector or valid path to a file openable with `sf::st_read`. Extent of content is used as the extent of the map. If numeric the value is used to scale a buffer around the aoi based on the aoi extent. Defaults to `1`.}

\item{watermask}{Either logical, sf-geometry, SpatVector or valid path to a file openable with `sf::st_read`. Polygons used to visualise water surfaces. If `TRUE` water polygons and lines are queried from OSM, if `FALSE` no water will be added. Defaults to `TRUE`.}

\item{aoi_colour}{Character, hex colour value without alpha. Defaults to `"#ff0000"`}

\item{landscape}{Either logical or `NULL`. Orientation of the map layout. Either forces landscape (`TRUE`), portrait (`FALSE`) or calculates it based on the orientation of the aoi (`NULL`).}

\item{aspectratio}{Numeric or `NULL`. The map extent is expanded to fit the ratio. If `NULL` the map extent isn't expanded at all. Defaults to `16/9`, the most common screen aspect ratio.}
}
\value{
Returns the final ggplot
}
\description{
Creates a overview map of a research area for use in papers and presentations.
While only the `aoi` has to be defined several other options for customization exist.
Since it was originally created for remote arctic environments the default visualization
focuses on topography and water bodies for orientation.
}
\details{
This function takes a polygon and creates a map for the area around it based on several parameters.
It supports all common vector data types. The function performs the following steps:
1. Load AOI.
2. Determine the map extent based on AOI or an additional vector geometry and a given aspect ratio.
3. Determine which country or region the AOI is situated in and load it's geometry for a inset map
4. Load custom basemap or calculate one automatically from a DEM. Downloads a DEM automatically if necessary.
5. Saves the final mosaic as a GeoTIFF file.
6. Load custom water mask or receive one from OSM
7. Create map with small inset overview map
}
\examples{
\dontrun{
# Create map with default settings for an existing sf object
map.aoi(researcharea_sf)

# Create fully customized map
map.aoi(
  aoi = "/path/to/aoi.shp",
  basemap = "/path/to/basemap.tiff",
  map_ext = "/path/to/mapextent.shp",
  watermask = FALSE,
  aoi_colour = "#0000ff"
  aspectratio = NULL
)
}

}
