% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_rem.R
\name{river.rem}
\alias{river.rem}
\title{Compute REM (Relative Elevation Model) from DEM and River Centerlines}
\usage{
river.rem(input_raster, centerlines, river_width, output_raster = " ")
}
\arguments{
\item{input_raster}{Path to the input DEM file in GeoTIFF format (`.tif`).}

\item{centerlines}{Either a `sf` object or a path to a shapefile containing river centerlines.}

\item{river_width}{Numeric. The width of the river network (in units of the DEM's projection), used for creating regular points along the river centerline.}

\item{output_raster}{Location the REM should be saved to. Defaults to the input_raster with a '_rem.tif' suffix.}
}
\value{
A `SpatRaster` object representing the Raster Elevation Model (REM) generated by subtracting the IDW interpolated raster from the original DEM.
The REM raster highlights areas of difference between the DEM and the interpolated values.
}
\description{
This function processes a high-resolution Digital Elevation Model (DEM) and existing river centerlines
to compute a Relative Elevation Model (REM). The REM highlights areas of elevation difference like paleochannels
by interpolating along the river centerlines and comparing the interpolated surface to the DEM.
}
\details{
The function performs the following steps:
1. Reads the input DEM and river centerlines.
2. Samples points along the river centerlines at regular intervals based on the specified river width.
3. Interpolates an elevation surface from the sampled points using Inverse Distance Weighting (IDW).
4. Resamples the interpolated surface to match the DEM resolution.
5. Computes the REM by subtracting the interpolated surface from the DEM.
6. Save the REM at the specified location

The function expects the river centerlines to be provided as an `sf` object or a shapefile with `LINESTRING` geometries.
If the geometries are `MULTILINESTRING`, they are automatically cast to `LINESTRING`.
}
\examples{
\dontrun{
# Compute REM from an input DEM and river centerlines
rem_raster <- river.rem(
  input_raster = "path_to_dem.tif",
  centerlines = "path_to_centerlines.shp",
  river_width = 50
  output_raster = "path_to_output.tif"
)
}
}
