% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filesort.R
\name{qgis.filesort}
\alias{qgis.filesort}
\title{Sort and Consolidate QGIS Project Files}
\usage{
qgis.filesort(input_path, output_dir = "", remove_old = FALSE)
}
\arguments{
\item{input_path}{Path to the .qgz QGIS project file.}

\item{output_dir}{Directory where all files will be copied. Defaults to a new folder
with the same name as the QGIS project in the QGIS project's directory.}

\item{remove_old}{Logical, if TRUE, deletes old files after copying.}
}
\description{
This function moves a QGIS project and all the source files for its layers
into a single directory. Source paths, styling, definition queries,
and metadata are kept intact. Layer files are sorted into appropriate
subdirectories ("Vector", "Raster", "Other") based on file extention.
}
\details{
This function processes a QGIS project file, copying all the associated
layer files and adjusting their paths to be relative to the new directory.
Layers are sorted based on their type.

Metadata, styling, and definition queries are preserved during the operation.
}
\examples{
 \dontrun{
qgis.filesort(input_path = "/path/to/project.qgz",
              output_dir = "",
              remove_old = FALSE)
}
}
