% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_rem.R
\name{river.rem}
\alias{river.rem}
\title{Creates Relative Elevation Models for Rivers}
\usage{
river.rem(
  input_raster,
  centerline = NULL,
  seed_point = NULL,
  output_dir,
  output_pattern = "riverdem",
  river_width = 10,
  overwrite = FALSE
)
}
\arguments{
\item{input_raster}{Either a `terra SpatRaster` or anything that can be read as one. DEM used to calculate the REM.}

\item{centerline}{Either a `terra SpatVector`, a `sf linestring` or anything that can be read as one. Centerline of a river. If not provided it will be calculated based on the seedpoint.}

\item{seed_point}{Either a `terra SpatVector`, a `sf point` or anything that can be read as one. Point in the riverbed from which the centerline will be calculated if no centerline was provided.}

\item{output_dir}{Path to save the output flow paths raster or vector file.}

\item{output_pattern}{Character. Will be used as part of the file names of any intermediate products.}

\item{river_width}{Numeric. Aproximate width of the river in meter. Used for point spacing, DEM smoothing and sink breaching.}

\item{overwrite}{Logical. Determines if the function is allowed to overwrite existing files.}
}
\value{
The REM as a `SpatRaster`
}
\description{
This function automates the workflow described in this QGIS Tutorial:
https://dancoecarto.com/creating-rems-in-qgis-the-idw-method
}
\details{
As input it takes a Digital Elevation Model and the centerline of the river.
If no centerline is provided it will be calculated automatically based on a user-defined seed point.
The seed point should be situated upstream inside the riverbed near the edge of the DEM.


The function performs the following steps:
1. Load the DEM and the centerline. If no centerline is provided, load the seed point.
2. If no centerline is provided, calculate one based on the seedpoint and the DEM.
2.1 Create a smoother version of the DEM using a median filter
2.2 Breach any sinks in the smoothed DEM
2.3 Fill any remaining sinks
2.4 Calculate Flow Direction
2.5 Trace downslope flowpath based on seed point
2.6 Turn raster flowpath into centerline vector geometry
3. Place points along centerline
4. Extract elevation values for these points
5. IDW interpolate the entire DEMs' area based on the extracted elevation values
6. Subtract the interpolated raster from the original DEM to get the REM
7. Save the REM
}
\examples{
\dontrun{
river.rem(input_raster = "path_to_dem.tif",
          centerline = "centerline.shp",
          output_dir = ".")

river.rem(input_raster = "path_to_dem.tif",
          seed_point = "seed_point.shp",
          output_dir = ".")
}

}
